Rename staking txns folder
